Rename n Modify File
Merubah nama file dari FXMLDocument.fxml menjadi NasabahForm.fxml
FXMLDocumentController.class menjadi NasabahFormController.class
