+++ Final Project/src/pkgfinal/project/FinalProject.java
@@ -15,9 +15,10 @@ public class FinalProject extends Application {
     
     @Override
     public void start(Stage stage) throws Exception {
-        Parent root = FXMLLoader.load(getClass().getResource("FXMLDocument.fxml"));
+        Parent root = FXMLLoader.load(getClass().getResource("NasabahForm.fxml"));
         
         Scene scene = new Scene(root);
+//        scene.setRoot(root);
         
         stage.setScene(scene);
         stage.show();
@@ -25,19 +26,20 @@ public class FinalProject extends Application {
 
     /**
      * @param args the command line arguments
+     * @throws java.lang.Exception
      */
-    public static void main(String[] args){
-        //        launch(args);
-        try {
-            Rekening rek = new Rekening(12, 100000.);
-            nasabahDataModel ndm = new nasabahDataModel("SQLITE");
-            Perusahaan temp = new Perusahaan("12435", 2, "Budi", "Kota Balam", rek);
-            ndm.addNasabah(temp);
-            System.out.println("Data berhasil ditambahkan");
-        } catch (SQLException ex) {
-            System.out.println("Gagal menambahkan!");
-            Logger.getLogger(FinalProject.class.getName()).log(Level.SEVERE, null, ex);
-        }
+    public static void main(String[] args)throws Exception{
+                launch(args);
+//        try {
+//            Rekening rek = new Rekening(12, 100000.);
+//            nasabahDataModel ndm = new nasabahDataModel("SQLITE");
+//            Perusahaan temp = new Perusahaan("12435", 2, "Budi", "Kota Balam", rek);
+//            ndm.addNasabah(temp);
+//            System.out.println("Data berhasil ditambahkan");
+//        } catch (SQLException ex) {
+//            System.out.println("Gagal menambahkan!");
+//            Logger.getLogger(FinalProject.class.getName()).log(Level.SEVERE, null, ex);
+//        }
          
     }
